Prion: DDP-4912, language popup infrastructure: refactored internationalization-related components into separate directory, cleaned up subscriptions in DisplayLanguagePopupServiceAgent, fixed issue where navigating to another component showed the language popup
