+++ ddp-workspace/projects/ddp-sdk/src/lib/services/serviceAgents/consentServiceAgent.service.ts
@@ -4,7 +4,7 @@ import { UserServiceAgent } from './userServiceAgent.service';
 import { LoggingService } from '../logging.service';
 import { ConfigurationService } from '../configuration.service';
 import { SessionMementoService } from '../sessionMemento.service';
-import { LanguageService } from '../languageService.service';
+import { LanguageService } from '../internationalization/languageService.service';
 import { Observable } from 'rxjs';
 
 @Injectable()
