+++ ddp-workspace/projects/ddp-sdk/src/lib/services/serviceAgents/displayLanguagePopupServiceAgent.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@angular/core';
+import { Inject, Injectable, OnDestroy } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { NotAuthenticatedServiceAgent } from './notAuthenticatedServiceAgent.service';
 import { ConfigurationService } from '../configuration.service';
@@ -7,12 +7,13 @@ import { StudyDetailServiceAgent } from './studyDetailServiceAgent.service';
 import { UserProfile } from '../../models/userProfile';
 import { UserProfileServiceAgent } from './userProfileServiceAgent.service';
 import { BehaviorSubject, Observable, zip } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, skip, take } from 'rxjs/operators';
+import { CompositeDisposable } from '../../compositeDisposable';
 
 @Injectable()
-export class DisplayLanguagePopupServiceAgent extends NotAuthenticatedServiceAgent<boolean> {
+export class DisplayLanguagePopupServiceAgent extends NotAuthenticatedServiceAgent<boolean> implements OnDestroy {
   private userDoNotDisplayObsSource: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-  private gotInitialProfileValue = false;
+  private readonly anchor: CompositeDisposable;
 
   constructor(
     @Inject('ddp.config') configuration: ConfigurationService,
@@ -21,6 +22,7 @@ export class DisplayLanguagePopupServiceAgent extends NotAuthenticatedServiceAge
     private profile: UserProfileServiceAgent,
     private studyDetailAgent: StudyDetailServiceAgent) {
     super(configuration, http, logger);
+    this.anchor = new CompositeDisposable();
   }
 
   private getStudyDisplayLanguagePopup(): Observable<boolean> {
@@ -28,19 +30,20 @@ export class DisplayLanguagePopupServiceAgent extends NotAuthenticatedServiceAge
   }
 
   private getUserNotDisplayLanguagePopup(): Observable<boolean> {
-    if (!this.gotInitialProfileValue) {
-      // Make sure we return the starting value from the profile
-      this.profile.profile.subscribe(res => this.userDoNotDisplayObsSource.next(res.profile.shouldSkipLanguagePopup));
-      this.gotInitialProfileValue = true;
-    }
-    return this.userDoNotDisplayObsSource;
+    // Make sure we return the starting value from the profile
+    const sub = this.profile.profile
+        .pipe(take(1))
+        .subscribe(res => this.userDoNotDisplayObsSource.next(res.profile.shouldSkipLanguagePopup));
+    this.anchor.addNew(sub);
+    return this.userDoNotDisplayObsSource.pipe(skip(1)); // Skip the initial false value
   }
 
   public setUserDoNotDisplayLanguagePopup(userDoNotDisplay: boolean): void {
     // Update the value in the profile
     const userProfile = new UserProfile();
     userProfile.shouldSkipLanguagePopup = userDoNotDisplay;
-    this.profile.updateProfile(userProfile).subscribe();
+    const sub = this.profile.updateProfile(userProfile).pipe(take(1)).subscribe();
+    this.anchor.addNew(sub);
 
     // Update our cached value
     this.userDoNotDisplayObsSource.next(userDoNotDisplay);
@@ -52,5 +55,11 @@ export class DisplayLanguagePopupServiceAgent extends NotAuthenticatedServiceAge
 
     return zip(studyDisplayObservable, userNotDisplayObservable)
       .pipe(map(([displayForStudy, doNotDisplayForUser]) => displayForStudy && !doNotDisplayForUser));
+  }
+
+  ngOnDestroy(): void {
+    if (this.anchor) {
+      this.anchor.removeAll();
+    }
   }
 }
