+++ ddp-workspace/projects/ddp-sdk/src/public-api.ts
@@ -72,7 +72,7 @@ export * from './lib/services/serviceAgents/temporaryUserServiceAgent.service';
 export * from './lib/services/renewSessionNotifier.service';
 export * from './lib/services/scriptLoader.service';
 export * from './lib/services/serviceAgents/governedParticipantsServiceAgent.service';
-export * from './lib/services/languageService.service';
+export * from './lib/services/internationalization/languageService.service';
 export * from './lib/services/serviceAgents/invitationServiceAgent.service';
 export * from './lib/services/serviceAgents/subjectInvitationServiceAgent.service';
 export * from './lib/services/serviceAgents/userInvitationServiceAgent.service';
@@ -85,7 +85,7 @@ export * from './lib/services/serviceAgents/submissionManager.service';
 export * from './lib/components/login/auth0-code-callback.component';
 export * from './lib/components/address/addressEmbedded.component';
 export * from './lib/components/activityForm/activity.component';
-export * from './lib/components/changeLanguageRedirect.component';
+export * from './lib/components/internationalization/changeLanguageRedirect.component';
 export * from './lib/components/popupWithCheckbox.component';
 export * from './lib/components/user/activities/userActivitiesDataSource';
